+++ mayan/apps/authentication/urls.py
@@ -1,43 +1,44 @@
 from __future__ import unicode_literals
 
-from django.conf import settings
 from django.conf.urls import url
-from django.contrib.auth.views import logout
 
 from .views import (
-    login_view, password_change_done, password_change_view,
-    password_reset_complete_view, password_reset_confirm_view,
-    password_reset_done_view, password_reset_view
+    MayanLoginView, MayanLogoutView, MayanPasswordChangeDoneView,
+    MayanPasswordChangeView, MayanPasswordResetCompleteView,
+    MayanPasswordResetConfirmView, MayanPasswordResetDoneView,
+    MayanPasswordResetView
 )
 
 
 urlpatterns = [
-    url(regex=r'^login/$', view=login_view, name='login_view'),
+    url(regex=r'^login/$', view=MayanLoginView.as_view(), name='login_view'),
     url(
-        regex=r'^password/change/done/$', view=password_change_done,
-        name='password_change_done'
+        regex=r'^logout/$', view=MayanLogoutView.as_view(), name='logout_view'
     ),
     url(
-        regex=r'^password/change/$', view=password_change_view,
+        regex=r'^password/change/done/$',
+        view=MayanPasswordChangeDoneView.as_view(), name='password_change_done'
+    ),
+    url(
+        regex=r'^password/change/$', view=MayanPasswordChangeView.as_view(),
         name='password_change_view'
     ),
     url(
-        regex=r'^logout/$', view=logout, kwargs={'next_page': settings.LOGIN_REDIRECT_URL},
-        name='logout_view'
-    ),
-    url(
-        regex=r'^password/reset/$', view=password_reset_view, name='password_reset_view'
-    ),
-    url(
-        regex=r'^password/reset/confirm/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
-        view=password_reset_confirm_view, name='password_reset_confirm_view'
-    ),
-    url(
-        regex=r'^password/reset/complete/$', view=password_reset_complete_view,
+        regex=r'^password/reset/complete/$',
+        view=MayanPasswordResetCompleteView.as_view(),
         name='password_reset_complete_view'
     ),
     url(
-        regex=r'^password/reset/done/$', view=password_reset_done_view,
+        regex=r'^password/reset/confirm/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
+        view=MayanPasswordResetConfirmView.as_view(),
+        name='password_reset_confirm_view'
+    ),
+    url(
+        regex=r'^password/reset/done/$', view=MayanPasswordResetDoneView.as_view(),
         name='password_reset_done_view'
     ),
+    url(
+        regex=r'^password/reset/$', view=MayanPasswordResetView.as_view(),
+        name='password_reset_view'
+    )
 ]